cdk.out/asset.98a9ad3d5bf9b766cf55e5e77277eb42917076d1ea5ac76872c3de75346010ef/pyproject.toml
[project]
name = "sudoku"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.212.0",
    "constructs>=10.4.2",
    "flask>=3.1.2",
    "mangum>=0.19.0",
]
